moved dijkstra to DRGEP, and created tools module for misc tools
